Update Source To vim-97.40.2
The original plan was to update the source to vim-91.100.1. Unfortunately, there were a lot of missing sources in that release.

+++ runtime/tools/ccfilter.c
@@ -85,7 +85,8 @@ const char USAGE[] =
 
 
 int ShowUsage( char *szError )
-{ int i;
+{
+  int i;
 
   fprintf( stderr, USAGE );
 
@@ -101,7 +102,8 @@ int ShowUsage( char *szError )
   return 0;
 }
 
-char *echogets(char *s, int echo) {
+char *echogets(char *s, int echo)
+{
  char * const retval=fgets(s, LINELENGTH, stdin);
  if (echo!=0 && retval!=NULL) {
   fputs(retval, stderr);
@@ -182,7 +184,7 @@ int main( int argc, char *argv[] )
 	  case COMPILER_GCC:
 	    Severity = 'e';
 #ifdef GOTO_FROM_WHERE_INCLUDED
-	    rv = sscanf( Line, "In file included from %[^:]:%u:",
+	    rv = sscanf( Line, "In file included from %[^:]:%lu:",
 			       FileName, &Row );
 	    if ( rv == 2 )
 	      {
@@ -191,11 +193,11 @@ int main( int argc, char *argv[] )
 	    else
 #endif
 	      {
-		if ((rv = sscanf( Line, "%[^:]:%u: warning: %[^\n]",
+		if ((rv = sscanf( Line, "%[^:]:%lu: warning: %[^\n]",
 				   FileName, &Row, Reason ))==3) {
 		 Severity = 'w';
 		} else {
-		rv = sscanf( Line, "%[^:]:%u: %[^\n]",
+		rv = sscanf( Line, "%[^:]:%lu: %[^\n]",
 				   FileName, &Row, Reason );
 		}
 		ok = ( rv == 3 );
@@ -203,24 +205,24 @@ int main( int argc, char *argv[] )
 	    Col = (dec_col ? 1 : 0 );
 	    break;
 	  case COMPILER_AIX:
-	    rv = sscanf( Line, "\"%[^\"]\", line %u.%u: %*s (%c) %[^\n]",
+	    rv = sscanf( Line, "\"%[^\"]\", line %lu.%lu: %*s (%c) %[^\n]",
 			       FileName, &Row, &Col, &Severity, Reason );
 	    ok = ( rv == 5 );
 	    break;
 	  case COMPILER_HPUX:
-	    rv = sscanf( Line, "cc: \"%[^\"]\", line %u: %c%*[^:]: %[^\n]",
+	    rv = sscanf( Line, "cc: \"%[^\"]\", line %lu: %c%*[^:]: %[^\n]",
 			       FileName, &Row, &Severity, Reason );
 	    ok = ( rv == 4 );
 	    Col = (dec_col ? 1 : 0 );
 	    break;
 	  case COMPILER_SOLARIS:
-	    rv = sscanf( Line, "\"%[^\"]\", line %u: warning: %[^\n]",
+	    rv = sscanf( Line, "\"%[^\"]\", line %lu: warning: %[^\n]",
 			       FileName, &Row, Reason );
 	    Severity = 'w';
 	    ok = ( rv == 3 );
 	    if ( rv != 3 )
 	      {
-		rv = sscanf( Line, "\"%[^\"]\", line %u: %[^\n]",
+		rv = sscanf( Line, "\"%[^\"]\", line %lu: %[^\n]",
 				   FileName, &Row, Reason );
 		Severity = 'e';
 		ok = ( rv == 3 );
@@ -228,18 +230,18 @@ int main( int argc, char *argv[] )
 	    Col = (dec_col ? 1 : 0 );
 	    break;
 	  case COMPILER_ATT:
-	    rv	 = sscanf( Line, "%c \"%[^\"]\",L%u/C%u%*[^:]:%[^\n]",
+	    rv	 = sscanf( Line, "%c \"%[^\"]\",L%lu/C%lu%*[^:]:%[^\n]",
 				 &Severity, FileName, &Row, &Col, Reason );
 	    ok = ( rv == 5 );
 
 	    if (rv != 5)
-	      { rv   = sscanf( Line, "%c \"%[^\"]\",L%u/C%u: %[^\n]",
+	      { rv   = sscanf( Line, "%c \"%[^\"]\",L%lu/C%lu: %[^\n]",
 				     &Severity, FileName, &Row, &Col, Reason );
 		ok = ( rv == 5 );
 	      }
 
 	    if (rv != 5)
-	      { rv  = sscanf( Line, "%c \"%[^\"]\",L%u: %[^\n]",
+	      { rv  = sscanf( Line, "%c \"%[^\"]\",L%lu: %[^\n]",
 				   &Severity, FileName, &Row, Reason );
 		ok = ( rv == 4 );
 		Col = (dec_col ? 1 : 0 );
@@ -270,10 +272,10 @@ int main( int argc, char *argv[] )
 		  }
 		 else
 		  {
-		    rv = sscanf( p+2, "%[^:]: %u: %[^\n]",
+		    rv = sscanf( p+2, "%[^:]: %lu: %[^\n]",
 				 FileName, &Row, Reason );
 		    if (rv != 3)
-		      rv = sscanf( p+2, "%[^,], line %u: %[^\n]",
+		      rv = sscanf( p+2, "%[^,], line %lu: %[^\n]",
 				   FileName, &Row, Reason );
 		    ok = ( rv == 3 );
 		  }
@@ -307,16 +309,16 @@ int main( int argc, char *argv[] )
 	  ok = sscanf( p, "make[%*d]: Entering directory `%[^']",
 		       BasePath );
 	  if (verbose)
-	    printf( "[%u]?%s\n", ok, Line );
+	    printf( "[%u]?%s\n", (unsigned)ok, Line );
 	}
        else
 	{
 	  for (p=Reason; (*p) && (isspace(*p)); p++);
 	  if ( BasePath[CWDlen] == 0 )
-	      printf( "%s:%u:%u:%c:%s\n", FileName, Row, Col, Severity, p );
+	      printf( "%s:%lu:%lu:%c:%s\n", FileName, Row, Col, Severity, p );
 	  else
 	    {
-	      printf( "%s/%s:%u:%u:%c:%s\n", &BasePath[CWDlen+1], FileName, Row, Col, Severity, p );
+	      printf( "%s/%s:%lu:%lu:%c:%s\n", &BasePath[CWDlen+1], FileName, Row, Col, Severity, p );
 	    }
 	}
       if (!prefetch)
